+++ src/main/kotlin/no/risc/encryption/models/EncryptionRequest.kt
@@ -1,7 +1,7 @@
 package no.risc.encryption.models
 
 import kotlinx.serialization.Serializable
-import no.risc.risc.models.SopsConfig
+import no.risc.crypto.sops.model.SopsConfig
 
 @Serializable
 data class EncryptionRequest(
